Moonsinger Holy Sites
Moonsinger Holy Sites have been added; Braavosa, Marconi, Licoria, Sud
and Fossati.

The Capital of the province 'Rosa' was set to a city instead of a
regular barony, due to being a coastal province in the republic of
Braavos.

A few tweaks to number of possible baronies in provinces were made. May
need to increase the average number of baronies in braavos in the
future, as the duchy capitals have capital + 3 sub baronies, which might
be a bit low.