Cities, Rhllor Holy Sites, Dragonstone and Driftmark
- Finished setting coastal provinces in Essos to cities.
- Tweaked some baronies here and there, set up Rhllor holy sites.
Dragonstone is now a holy site of the Rhllor faith, but the rest of the
holy sites can be found in the free cities. In short, they are currently
linked to significant 'faith' cities.
- Made tweaks to the provinceDef.xls file itself. Cleared up some blank
rows, coloured the uncoloured Kingdom rows to make the file more
readable. Hopefully, as a few of the colours are borderline too
contrasting for me...
- Massey's Hook is no longer crownlander, but the superior Stormlander
culture.
- Get ready for build!