Dynasty and Character File Resort
Vierwood had changed the names, but did not put the ID ranges inside the
file through the use of a comment, so I went ahead and did it for all of
the files. I've also added character files for every single region with
pre-assigned ID ranges.

At the top of each file the ID range to use for characters is in there.
We can add additional ranges to the file if we get to the point in which
we have more than two hundred unique characters per region.

The name format uses no spaces or brackets, to assure that the mod works
on all systems and to make sure the game has no issue reading files.

I also edited some of the files to fit the general format scheme better.
If you've made changes to these files, it should be quick and easy to
merge the changes in. I also fixed one of the north minor house
characters, who had a conflicting ID with another (600) to fit into the
North bracket.