New Religions, Tweaks
Main changes included; Trios, Boash'i, the Bearded priests. The former
two religions have holy sites assigned, and are present in starting
provinces. The bearded priests should be added to Lorath down the line,
scattered over a number of provinces.

A Rhllor holy site was moved to Lorath from the red temple in Myr.