Valyria Fixes, Other Edits
Valyrian Empire no longer exists whilst an attempt to fix the Valyrian
Freehold is underway. The characters work, but the republic itself is
broken however.

A republic may not work with republics under it, as it forces the
various King-Tier republic rulers into the Valyrian Empire-Tier republic
as houses that can fight for the republic, instead of the assigned
count-level houses for the Valyrian Republic.