Mod Build, Tweaks
Lys (the islands) have been buffed somewhat. This is due to a Valyrian
Kingdom being ruled over the island, and because of the present five
houses there. Balance wise, it should be a good because it helps protect
them against their vassals.

From what I  (Rowan) am aware of, Lys was already well built and a
hot-spot for the Valyrians so it may make sense from a lore perspective.

Valyrian religion holy sites added to every Lys province. For now so
they have full control over their faith, but in the future the sights
can be spread out (into westeros, even).