Wolfswood
Tanner suggested it, and having looked at the de jure blob that is
Winterfell, I feel it might be the best course. Plus, Winterfell and
Moat Cailin are two of the best duchies in the North (or will be).

That, and the Kingdom of Winter / Winterfell will be easier to make with
only two required duchies.