Building Rebalance
- Reduced fort level and tax income overall for the core castle
buildings.
- Reduced the overall levy given by the core buildings, but made the
units given of a wider variety with a focus on heavy infantry.
- Reduced the impact on tax per castle core level, but also reduced the
bonus given by the administration tree. The bonus in the administration
tree now grows with further development, helping counter the negative
effects of other buildings and then give a small bonus later on.
- Increased the tax given by the higher level castle town chain
buildings and reduced their impact on local tax income.
- Recruitment post now gives more levies.
- Walls now give slightly less fort level per level.