Province Edits, Unique Buildings Cleanup
- Secourt, the Eyrie, renamed to the Mountain of the Moon and given more
barony slots (5). Gates of the Moon was given 6 barony slots instead of
seven. May need rebalancing?
- The Marcher castle building now works and will be given on start, as
the potential to build was not in an OR statement, but outside. So the
province needed to be of a lot of provinces to be valid. In other words,
could never be valid!
- Most unique buildings will be given on start.
- A number of buildings that had provinces manually assigned will now
generate based on culture group as a requirement (iron isles and shield
isles, for example).
- Ordering is better, with spacing to seperate costs, bonuses, maluses
and so on. Makes it all more readable.
- Moat Cailin can now only be built (and actually exists now) in the
capital castle.
- Nerfed wildlings hard, so that one barracks building will give a few
hundred men, rather than a few thousand.
- Removed OR statements where there was only one possible option.
- Made a statement at the top of the file about additive - seems a
number of the upgrades were made without taking into account previous
ones, such as with the wildlings, which meant thousands of men coming
from a single holding!
- The religion of the province instead of the current holder is taken
into account for weirwoods. Changing the religion of the province will,
however, still destroy the building.
- Kingswood and Rainwood resources moved into the unique resource file
and disabled, as I added buildings a while ago that handle this, taking
advantage of the resourece allocation event.
- Arbor, Driftmark, Lannisport and Oldtown shipyards fixed.
- Added resources to more locations. Mostly the Vale and Westerlands at
the moment. Rainwood, Mistwood, Wolfswood and Lemonwood all added to the
map.
- Lots of other little changes...