Triggered modifiers replacement
Triggered modifiers should be avoided like the plague, they account for a huge amount of the games slow down.
They are modifiers checked against every single ruler, every single day for every single modifiers. The fact they are even a thing in the game is a crime against coding.
Just the single vanilla one is an awful thing to have.
Instead add them as characters modifiers and use the modifier maintenance events I have added to assign them, it is only checked on startup, yearly and upon gaining a title but it is far more efficient and there is really no situation the normal triggered modifiers is a better idea to be used.