Small Batch of 'Fixes'
We have some major issues... To say the least.

I've begun fixing some of them, but for the most part we have a
widespread problem of missing events that are still being called and so
on.

I have re-added a lot of vanilla events. We need to go through them one
at a time, remove references to said event or edit them to be compat.