Reach Balancing
The Reach is now heavily decentralised, with all of it's power in the
hands of the marshals. As a result, if they are on the council the Reach
will be very stable.

Howerver the Marshals must agree on most matters and may not
independently wage war.

I have included a new 'test' law that may or may not prevent war outside
of the realm but still allows internal conflict. It may not work.