Tweaks
Changed the stats of infantry a lot, heavily inspired by the changes
made by EMF for HIP. Not exactly the same, due to our light infantry
being massed unwanted levy and all that.

Also reduced maintenance costs overall. They appear to be bad before and
perhaps still because most are used to modded values, in which levies
cost very little.

Furthermore, many of the raised troops are in personal demense rather
than vassals, so you are personally paying for most of them!