Major Nerf
Knights are by far the best unit in the game. A single knight is worth
many, many more troops. Buffing Highgarden with additional units is not
good at all - it is turning the most powerful building already into an
absurdly powerful structure and further buffing the most powerful
Kingdom in the game (no other Kingdom will ever be able to rival them in
levy or strength at this rate).

I balanced all of the uniques to be in-line with one another, based on
what they offer.

Buff his holdings with more development.