Sadface
It deleted my faction changes. Whoo!

~ Buffed the base value of cities, castles and temples for a number of
reasons. Should also help make raiding more worthwhile, as there will be
more to raid.
~ Extensive number of custom scripted triggers, used to enhance factions
- Dornish are now less likely to join factions against their liege so
long as their liege is Dornish.
- Marcher cultures are now less likely to join factions against their
liege so long as their liege is of the same group.
- Old Gods followers are more loyal to their own people.
- Drowned, Valyrian, Seven and Eastern Pagans are far less loyal when
under a foreign religion.
- Valyrians (culture) are just more ambitious.
- Council members are more loyal, even with no empowered council.
However, they are also more dangerous if they do not like you (and are
not empowered).
- Eastern Religion group and cultures are generally less likely to form
factions.
~ Dukes may now form or join the loyalist faction.
~ Moved reachman, manderish and shield islander into first man group.