Subjugate Test (Needs Testing)
Quick redo, so you subjugate the entire (target) realm rather than just
the target title. Waging war against a King will require the use of
Duchy and/or County Conquest now.

- Should be able to use against Kings, not Empires.
- Should take the King and ALL Kingdoms they hold as a vassal without
stripping titles. Before the King would remain independent as a count if
they held a province directly or as a vassal outside of the target
Kingdom.
- AI can also now use the CB, but I've restricted it a bit to tone it
down. Needs testing.