Potentially Fixed Potential
Vanilla requires that you are neither Muslim or Indian to have the
elective law, unless you already have elective or voting laws for it.
This was resulting in some funny business, because though the law would
be set for the Ironborn, it would revert on title change.

Probably due to a vanilla conflict, because it's not possible to have
the law at all in vanilla for some reason, if you're only checking for
law already OR voting power, and that may have been causing the default
in gameplay.

*fingers crossed*