More Tweaks
~ No longer religion restricted. Instead a yearly event will fire if you
have a knight trait, but aren't of the correct religion. As a result,
you'll lose it and get hedge instead.
~ Minor title for Ser, may or may not work. Pretty sure minor title is
what comes first, nickname second.
~ Tweaked the event heavily, so hedge_knight is a fallback trait from
the chain.
~ Can name page, squire and upgrade to the knights at reduce cost. All
of this requires that you are a Knight. To become a Knight, you must be
knighted by one or you must request the order to knight you.