Resource Allocation Test, Rewrite
Initial test draft. May not work because I am a derp.

Using trigger_switch and firing on a per-character basis rather than for
a single character, on suggestion of Zara.

If it works OK, it should open for some degree of random assignment of
modifiers and other fancy business without the previously encountered
desyncs. It also scopes per region to hopefully reduce impact.