Update
- Tweaked, reduced number of hard province ID references. Some modifiers
are given to entire duchies now, given the correct local terrain.
Wolfswood will, for example, be given to any province in that region so
long as it has forest or woodland terrain.
- Arid will be assigned to every single province in Essos and Dorne
(Excluding deserts in Dorne, since they have an equal chance of being
lush versus bleak). Arid may be turned into a hidden modifier.