Time to Build Consideration
The time taken to build is now also considered for the overall time to
pay back an entire tree up to x level. This means the player in
question, for the calculation to be accurate, is building each upgrade
as quickly as possible - otherwise it will take longer than expected.

Example: If you upgrade the lumbermill to level 1, it will take 127
years to pay off. If you wait 127 years before upgrading, you will be
incredibly inefficient. If you upgrade as quickly as possible to level
5, you can pay off the entire tree in around 60 years.