Maximum Years to Cover (YtC)
Calculation for maximum years to cover the costs of constructing a
particular tree. An overestimate at worst, it takes into account the
maximum time to cover each upgrade as it comes. Lumbermill, for example,
will take an absolute maximum of 368 years to completely cover the cost
or a minimum of 58 years.*

* This does not take into account local build cost reductions, and local
accumulated tax modifiers that may boost the structure's income.
Furthermore, it does not consider the affects of your steward on build
time/build cost/local tax modifier either.