More Nicknames
Talked to the guy who made Nicknames ++ and he said it was ok that we could use his nicknames in our mod.

I also talked to the guy who makes all of those VS mods and he said we could use his birth control mod, so the option is there if players want less children.