Tweaks and more
Myrlanda Pyper is now the mother of Robert Vance. Roberts age is set to 1.
Any mountain is now a province making it possible to use the "Realm-color Wastelands" feature. I also added a missing lake.
The King of the Iron Islands and the King Beyond the Wall can now be proclaimed via ambition.