Flickering map fix and province tweaks
The flikering map at the disputed lakes and other region is fixed. It was caused by the topology map ( color was 95 95 95, so the engine could not "decide" between sea or land.
Minor islands are assigned to nearby provinces.
Many many small tweaks.
Would be great if someone could make a new height map based on the new topology map. It works fine so far, but to avoid any issues making a new one is necessary.