Weeping Water and Green Fork/Twins
The Weeping Water beside the Dreadfort has been added, as well as an extension of the Greek Fork of the Trident upriver of The Twins.

There is now an impassable barrier following the river system, all the way to the western coast. I have added more marsh to indicate that the terrain there is too rough for an army, and also provided a bit of farmland to Niceland and Mistlewood.

Provinces bordering the river system have been altered in shape. The Frey Woods should be swapped from the Mistlewood duchy to The Twins duchy by Vierwood, because Frey Woods no longer borders the rest of the Mistlewood provinces.

Rowan could also add a crossing between Mistlewood province and Niceland if desired, like Moat Cailin and Lowbarrow. This would reduce importance of The Twins, but an alternative western route is plausible and even probable in lore.