Overhauled portrait backgrounds; tweaks
New portrait properties almost completely remade highly based on CPRplus.
New backgrounds are added from CPRplus and the Elder Kings (with permission).
Old Gods priest have now the pleasure to face the weirdtree's face (weirwoodtree background). I have also added a snowy forest. There is a possible catacomb for the faceless men. I will activate it if there as soon as a society is added for them.
The Orphans of the Greenblood are now Rhyonar. This is more lore-friendly.
I added missing localisation for the invader cultures (Dothraki, Ghiscari, Rhoynar