Red Lake + Normal Height Map matched to Topology
I've added Red Lake (the actual lake) inside Red Lake province and adjusted the map position of the town, situating it to the NE of the lake in compliance with lore. More farmland and green bog now surrounds the lake.

A new normal height map has also been generated to match TitanRogue's topology map. I placed a backup of the old topology and height map files in my personal subfolder for future experimentation on them.