More adjusted province history
I also updated the provinceDef file. It now has the culture and religion values of 1.1.1. So Westeros is First Men and Old Gods only.
Like the previous commitment these edits aim to make bookmark and history modding for provinces easier and faster. I will continue with the rest in future time.