landing occluded chars
I know we didn't want them at first, but having these chars ruling titles provides more playable history and does NOT mess with the regnal numbering, because they are named as "Line of..." not the actual char's name. Please don't remove these.

This adds in the Yronwood line of Yorick II, the Justman line of Lucifer I, the Lemonwood line of Gerris, and the line of Clarence and Rupert Crabb.