Argos Invasion of the North
Historically Rogar Bolton submitted to the Winter King Theon Stark when Andals invaded the North. They defeated the leading Andal warlord Argos in the battle of Weeping Water.
House Bolton is now a tributary of the Starks. Argos Sevenstar has got some new units added upon start-up to be a threat to the First Men.