Dragon Engine | Update
More duel work;
You should now lose your old trait when leveling your dragon riding skills;
You now get an encounter when participating in a duel. For now it is team A who leads every attack but later on the winner of the first round will lead the second attack. I also want to make the outcome of duels heavily dependent of the dragon strength as for now it is evaluated randomly;
Fixes and Tweaks;
The event pictures that are added will not be visible;