Fixed some capitals
Some capitals were not set correctly so when starting a new game the capital of the high lordship of Deep Den was not Deep Den but something else which made no sense.
I made no changes to the kanded_titles file. I switched the positions of duchies in provinceDef so that the capitals were on top of its duchy and than ran the mapfiller tool.